pyproject.toml
[tool.poetry]
name = "flask-auth-demo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "flask_auth_demo"}]

[tool.poetry.dependencies]
python = "^3.8"
PyJWT = "^2.8.0"
Flask = "^2.3.2"
bcrypt = "^4.0.1"
SQLAlchemy = "^2.0.19"
python-dotenv = "^1.0.0"
uuid = "^1.30"
psycopg2-binary = "^2.9.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
